Fixing includes within Calypso for the move of shipping classes from woocommerce-services to woocommerce in the state.
