+++ client/extensions/woocommerce/state/action-types.js
@@ -332,3 +332,6 @@ export const WOOCOMMERCE_UI_REVIEW_REPLIES_EDIT = 'WOOCOMMERCE_UI_REVIEW_REPLIES
 export const WOOCOMMERCE_WC_API_SUCCESS = 'WOOCOMMERCE_WC_API_SUCCESS';
 export const WOOCOMMERCE_WC_API_UNAVAILABLE = 'WOOCOMMERCE_WC_API_UNAVAILABLE';
 export const WOOCOMMERCE_WC_API_UNKNOWN_ERROR = 'WOOCOMMERCE_WC_API_UNKNOWN_ERROR';
+export const WOOCOMMERCE_SHIPPING_CLASSES_REQUEST = 'WOOCOMMERCE_SHIPPING_CLASSES_REQUEST';
+export const WOOCOMMERCE_SHIPPING_CLASSES_REQUEST_SUCCESS =
+	'WOOCOMMERCE_SHIPPING_CLASSES_REQUEST_SUCCESS';
