+++ client/extensions/woocommerce/state/sites/shipping-classes/actions.js
@@ -6,14 +6,14 @@
 
 import { errorNotice } from 'state/notices/actions';
 import {
-	WOOCOMMERCE_SERVICES_SHIPPING_CLASSES_REQUEST,
-	WOOCOMMERCE_SERVICES_SHIPPING_CLASSES_REQUEST_SUCCESS,
-} from 'woocommerce/woocommerce-services/state/action-types';
+	WOOCOMMERCE_SHIPPING_CLASSES_REQUEST,
+	WOOCOMMERCE_SHIPPING_CLASSES_REQUEST_SUCCESS,
+} from 'woocommerce/state/action-types';
 import { areShippingClassesLoaded, areShippingClassesLoading } from './selectors';
 
 export const fetchShippingClassesSuccess = ( siteId, data ) => {
 	return {
-		type: WOOCOMMERCE_SERVICES_SHIPPING_CLASSES_REQUEST_SUCCESS,
+		type: WOOCOMMERCE_SHIPPING_CLASSES_REQUEST_SUCCESS,
 		siteId,
 		data,
 	};
@@ -26,7 +26,7 @@ export const fetchShippingClassesFailure = () => {
 };
 
 export const fetchShippingClassesIfNotLoaded = siteId => ( {
-	type: WOOCOMMERCE_SERVICES_SHIPPING_CLASSES_REQUEST,
+	type: WOOCOMMERCE_SHIPPING_CLASSES_REQUEST,
 	siteId,
 } );
