+++ client/extensions/woocommerce/state/sites/shipping-classes/test/data/initial-state.js
