+++ client/extensions/woocommerce/state/sites/shipping-classes/test/selectors.js
