+++ client/extensions/woocommerce/woocommerce-services/state/action-types.js
@@ -183,7 +183,3 @@ export const WOOCOMMERCE_SERVICES_SHIPPING_ZONE_METHOD_SETTINGS_REQUEST_SUCCESS 
 	'WOOCOMMERCE_SERVICES_SHIPPING_ZONE_METHOD_SETTINGS_REQUEST_SUCCESS';
 export const WOOCOMMERCE_SERVICES_SHIPPING_ZONE_METHOD_UPDATE =
 	'WOOCOMMERCE_SERVICES_SHIPPING_ZONE_METHOD_UPDATE';
-export const WOOCOMMERCE_SERVICES_SHIPPING_CLASSES_REQUEST =
-	'WOOCOMMERCE_SERVICES_SHIPPING_CLASSES_REQUEST';
-export const WOOCOMMERCE_SERVICES_SHIPPING_CLASSES_REQUEST_SUCCESS =
-	'WOOCOMMERCE_SERVICES_SHIPPING_CLASSES_REQUEST_SUCCESS';
